requirements.txt
langchain
langchain-classic
langchain-community
langchain-chroma
langchain-huggingface
chromadb
tiktoken
pypdf
sentence-transformers
transformers
huggingface-hub
python-dotenv
streamlit
langchain-groq
watchdog
pymupdf